Insert split command at appropriate time
fix bug where windwow incorrectly thinks a window arrangement has
been reconstructed. before, windwow would only check if the current
window matched when inserting switch commands into an existing window
command list. this caused a bug when there were multiple windows with
the same dimensions - thus resulting in a false positive.

this was fixed by adding a second recursion step to the matching case
and checking to see if the final list of windows is the 'same' as the
one originally given.